VArray to Dict casts should not raise varray promotion notices
Summary:
A varray to dict cast was using the same machinery as a varray to
darray cast. This meant, as an unintentional side-effect, a varray promotion
notice was being raised when those notices are enabled. Since this is just an
implementation detail, provide a parameter to the relevant functions to suppress
the notice for this case. This isn't just cosmetic. In debug builds this will
crash the VM because we don't sync the VM registers for dict casts.

Reviewed By: paulbiss

Differential Revision: D9675827

fbshipit-source-id: ab9f054605b4959fe5a60bdbb0838210362489b1