Handle const declarator lists the same way as property declarator lists
Summary:
Currently, multiline const initializers like this one:

```
  const vec<string> SYMMETRY = vec[
    'Tiger, tiger, burning bright',
    'In the forests of the night',
    'What immortal hand or eye',
    'Could frame thy fearful symmetry?',
  ];
```

Are formatted like this:

```
  const vec<string>
    SYMMETRY = vec[
      'Tiger, tiger, burning bright',
      'In the forests of the night',
      'What immortal hand or eye',
      'Could frame thy fearful symmetry?',
    ];
```

This is the appropriate behavior when there is more than one item in the declarator list:

```
  const vec<string>
    FOO = vec[],
    SYMMETRY = vec[
      'Tiger, tiger, burning bright',
      'In the forests of the night',
      'What immortal hand or eye',
      'Could frame thy fearful symmetry?',
    ];
```

But is unnecessary when there is only a single item whose name fits on one line--in that case, we want the split to be independent rather than parental. We already have handling for this behavior (used in property declarator lists; see test case static_initialization.php), we just need to make use of it in const declarations.

Reviewed By: arxanas

Differential Revision: D9672424

fbshipit-source-id: c5a33020a60e6034ff9f007b70e4fb19d67df76d