Add HH\is_list_like
Summary:
Add HH\is_list_like. This new builtin provides an efficient way to
check if the input is an array-like which is "list-like". This means it only
contains of integer keys from 0 to N-1, in that order. Vecs and varrays are
always list-like, while other array types may be.

Reviewed By: mxw, alexeyt

Differential Revision: D9687748

fbshipit-source-id: 58ba57cd45ded8d511c3baece11357369385b39c