Test for Errors.incremental_update ordering
Summary:
Quoting from `errors_test.ml`:

"Tests documenting various invariants about the order of things coming out of Errors module. Some of them are __probably__ not necessary for correctness, but if you break them you will still change a bunch of tests outputs and will have to spend time wondering whether the changes are significant or not."

This is exactly what happens in {D8384219}, so I'm extending the tests to have more than one error per file, to show how errors are reversed or not by `incremental_update` method.

Reviewed By: arxanas

Differential Revision: D8392466

fbshipit-source-id: 0938e5a606c6f842323e33514e343408d584f086