Pre-declare files before processing didOpen/didEdit/didClose messages to reduce fan-outs
Summary:
One of big problems with incremental mode is the fact that it only works well when we know what was the "previous" version of declarations that have changed. With saved state, lazy init and lazy parsing we have moved away from storing all of them, so incremental mode often has to give up and instead use a very conservative full fan-out.

Summoning previous version of declaration from the past is a hard problem in general, because declarations depend on global state - what `class Foo` means depends on naming table telling you where Foo is defined (and in many cases, it's inheritance ancestors), so we not only need to know the previous versions of the files (which in itself requires an expensive `hg cat`), but also previous version of entire naming table, which at that point is likely irreversibly gone, replaced by current version.

While the general case needs a general solution (that will likely require some big rearchitecting of incremental mode), there are special cases that can be solved much more easily - when editor sends `didOpen/didClose/didChange` notification, we know for sure at least one file whose declarations we'll need, and we know it BEFORE anything changes (since we are in control when to apply the actual modification). We have both current naming table and current file version, that soon will become the "previous" one.

Is it going to help anything in practice? I add config flag and logging to be able to A/B test. Anecdotally, I was going through the samples of big rechecks where prechecked files were enabled (D9620964) and were not enough to help, and I got the impression that a big part of them were cases of people just BROWSING some core files in editor[1], likely through go-to-def, yet they end up with some horrible fan-outs. We'll see.

The only risk I see here is that this additional declaration in a place that wasn't there before increases chances of some race conditions like the one in D9601566. But since everything is already broken...

Reviewed By: arxanas

Differential Revision: D9666691

fbshipit-source-id: 27361db276b3cc8c3af35b7165b15d2682744e29