Move the checks performed when type-checking switch statements into a linter
Summary:
At runtime, the scrutinee of a switch statement is compared to case values using the loose `==` comparison.  We don't enforce any relationship between the types of the arguments of `==` elsewhere in Hack.  (In fact, we don't do this even for strict `===` comparisons.)  That's why I'm suggesting to remove any checks performed when type-checking switch statements out of the type-checker and move them into a linter instead.

At the moment, we check if the scrutinee and case value are both subtypes of `num` or `arraykey`, and if not we check if their types unify (with the intention to enforce that the types must be the same).  I'm not sure how useful these checks are, so I'm proposing an alternative: if the type of the scrutinee is fully resolved, check that the types of the case values are subtypes of the type of the scrutinee.  The intuition is that the case values will typically be constant expressions whose types will be statically known.  So, if we know, say, that the scrutinee of a switch statement will always be an `int` and one of the case values is a string, then probably we are doing something wrong (rely on loose comparison, or have overlooked that the string case is redundant etc.).  This also covers the newtype examples from [[ https://fb.facebook.com/groups/472626396119314/permalink/2025783637470241/ | my post ]] because subtyping of newtypes with the same name disregards the bounds and checks only the generic parameters (if any).

Reviewed By: Wilfred

Differential Revision: D9150088

fbshipit-source-id: 540cea10ddc381948389ab5bac9d46097feb0f60