Remove unnecessary mask in hh_header_t construction
Summary:
The header bytes only provide 31 bits for storing both the size and uncompressed size. To protect against something larger being stored, I originally masked the number before combining it with the rest of the header bits.

In the event that the mask actually did anything, it would cause exceptions downstream as the stored uncompressed size would not match the actual uncompressed size. Either LZ4 decompression would fail or the assert in that function would throw.

Happily, local invariants conspire to ensure that both size and uncompressed size will fit. Before compression, we ensure that the original size fits in 31 bits. We also ensure that the compressed size is less than the original size, so both size and uncompressed size will fit in 31 bits.

Since both numbers will certainly fit in 31 bits, we don't need the mask.

Reviewed By: gabelevi

Differential Revision: D9683818

fbshipit-source-id: 0ff6360a89546b0ace10fd3125749965475df159