Find refs support for prechecked files
Summary:
When using prechecked files, dependency table can be incomplete. Find refs and refactor use it, and need special handling for it. I add uniform treatment of all symbol kinds, but couldn't come up with a test case for anything but class members. Meaning of "bar" in "X->bar" can change based on related changes in other files (that can influence type of X), but I can't see how a function/class/constant reference would suddenly appear in the same way in a file that it was not in before.

(Note: this ignores all push blocking failures!)

Reviewed By: 2BitSalute

Differential Revision: D9697286

fbshipit-source-id: 20dcb88ae2b31b98e1548c8957fa36903f065468