Emit CallToExit stub using vasm, and move enterTCExit to hot
Summary:
This does not bypass the call to exit stub, but it should allow it to be a single direct jump.  There was no reason it was being emitted using the assembler rather than vasm.

Completly bypassing callTOExit by having it alias enterTCExit will require changes that allow us to still differentiate between callsites.  This is a little more involved, so for now do the easy thing.

Depends on D9660642

Reviewed By: jjgriego

Differential Revision: D9750593

fbshipit-source-id: 8b56b668c43245b47b1e8e486923036cd1735bd6