Let relocator change sequences using rAsm
Summary:
Currently the relocator only allows undefing registers if they are vixl scratch registers. This is done so that it doesn't undef a live register.
This adds the vasm scratch register to the set of registers it can undef.  This should be particularly useful since calls to helpers, and some jumps use the vasm scratch register to hold the address they will branch to.

Reviewed By: ottoni, jjgriego

Differential Revision: D9750612

fbshipit-source-id: a34ae948c589c926decd7799d965ff95bce0b41d