Tests for Dep.AllMembers effectiveness on reducing fan-outs
Summary:
This was added in D6469069, as an example of improvements we can make with more fine-grained dependencies.

There was an integration test for correctness, "real life" test for effectiveness[1], and some maybe-what-ifs about saved state scenarios.

As we rely on saved state more and more, I think it's time to revisit the last part, and this test shows the problems we need to solve. Some investigation of fan-outs (T33848406) that I started doing as part of pre-checked files roll-out shows that it might be valuable to special case several "core" (mostly enormous autogenerated enums) definitions in saved state (T33848864).

Differential Revision: D9780075

fbshipit-source-id: 6fc62dc07f48d56ce05df84f0778d8f01e97e2f9