Add use of disjunction to subtype solver
Summary:
In future we will generate disjunctions in the subtype propositions that are returned by `simplify_subtype`.
Here is a simple case; we currently "solve" disjunctions using the hacky `Errors.try` as used by the legacy subtype
checker, but in future we will reason about disjunctions in a more functional style.

Reviewed By: CatherineGasnier

Differential Revision: D9752474

fbshipit-source-id: 55dccfca2ccf8e5d0b9c41098334ba00688835e6