Experimental Xenon PerRequest Mode
Summary:
In some cases, having a per-request sampling gives us the benefit of knowingly changing the behavior of the request despite the fact we know we are skewing the measurement.
The new option of RequestFreq allows us to do by extending Xenon's current behavior (while not altering it).
A frequency of 0 will disable Xenon logging and a frequency of 1 is practically the current state (always process xenon interrupts).

A frequency higher than 1 will allow us to turn Xenon once per many requests with a highly granular period.
Requests will be able to query whether Xenon is on and behave differently.
This will benefit cases in which extra contextful stack information (e.g. registration stacks of action delegates) can be collected during the request that is Xenon profiled.

Reviewed By: jano

Differential Revision: D9757668

fbshipit-source-id: 93925113085401ca3e6cb2ca1262f4ef56ab395c