un-inline VectorSet
Summary:
Building on top of the instrumentation in D9392699 showed that the inlined
code for VectorSet only executes 1/1000 to 1/100 as often as array set, both
for the copy and in-place cases. In many cases, the in-place (fast) code path
was only slightly hotter than the copy codepath, undermining the value of having
an inline fast path.

It is likely that the inlined code for VectorSet was added to reduce the overhead
of Vector compared to PackedArray. Now that we're migrating to vec, the Vector
optimization probably doesn't make sense anymore, and complicates adding
tracepoints for copy-on-write.

Reviewed By: alexeyt

Differential Revision: D9683467

fbshipit-source-id: 90ccfd07bfcb7a945c4471817c56e9d3d3725403