Kill Nast_terminality
Summary:
Following the work on typing with continuation, this diff removes all uses of Nast_terminality.

To do so, `nastInitCheck`, which checks that all members are initialized in a constructor, needs to be more clever with control flow.

For example, if we have an `if` statement like:
```
if (...) {
  // This branch initialize a set of properties S1
  ...
} else {
  // This branch initialize another set of properties S2
  ...
}
```
then the set `S` of properties initialized by this `if` statement is the intersection `S1 /\ S2`.

If one of the branches throws, say the first branch, then the set `S` of properties initialized by the `if` statement is equal to the set of properties initialized by the branch that does not throw, i.e. `S = S2`. This amounts to saying that `S1` is some top element of the set of sets of variables, which we call `Top`, which has the property that for all set S of properties, S is included in `Top`, such that `S = S1 /\ S2` still holds.

Reviewed By: manzyuk

Differential Revision: D7743824

fbshipit-source-id: 2aede74c081668670b7878ba4b9a4a6cb868c1ff