union is not unification
Summary:
So far, the way to perform a union of two types was to wrap each in a `Tunresolved`, create a fresh `Tvar` pointing to those `Tunresolved` then perform a unification on them.

Although this would create the desired result in most cases and avoid duplication of code, semantically, a union is not a unification, and there are other ways to unify than performing a union which we might want to use in a near future (e.g., constraint-based unification).

The main drawback of this approach is the unnecessary use of `Tvar` and `Tunresoveld` wrappers, which remain in the resulting union type. This is problematic in some cases, as the result union type is actually not a type variable and should not be allowed to "grow", as `Tunresolveld` do, further down the code.

Besides, the wrapping in `Tunresolved` would silence a few errors like using `?->` on nonnull things, since `?->` is allowed on any `Tunresolved`. A lot of errors revealed by this diff in www are of this sort. See also test hphp/hack/test/typecheck/union/null.php

There are a few other subtle differences between a proper union as we write it here and unification:
* `Tvar`: unioning two `Tvar`s should actually create a third independant `Tvar` pointing to the union of the types pointed to by the two first `Tvar`s, whereas as a result of unification, all three `Tvar`s will point to the same type. This would obviously not work when dealing with control flow with continuations. (This allows us to get rid of the rather ugly `make_union_instead_of_unificaton` flag in `Typing_unify` which we had to resort to in the previous diff of this stack to make typing with continuations work).
* `mixed` unions with anything as `mixed`, `nonnull` unions with any option as `mixed`, and with anything else as `nonnull`, which wasn't the case in the unification code.
* shapes: The code for shapes was quite complex and wasn't performing what we expected in in some cases. For example, in
```
$myshape = shape();
if (some_cond()) {
  $myshape['a'] = "something";
}
```
we want the resulting type of `$myshape` after this `if` to be `shape(?'a' => string)`. The unification code would produce `Tunresolved(shape() | shape('a' => string)`, which would cause code like the one in `hphp/hack/test/typecheck/union/shape.php` to produced a `Tunresolved` of exponentially growing size. The complexity of the existing code for shape unification, which would perform a lot of checks and raise many errors, made it hard to fix for this case.

The resulting code for the union is overall simpler than the unification code, because contrary to unification, union always succeed. Therefore, there is for example no need to perform various checks and create errors.

As we are working on better ways to perform unification, this is a first step in preparation for this future work. We are now free to experiment with unification without fear of breaking union. Besides, here, we only create `Tunresolved` (which should really be `Tunion`) when necessary, which is a step forward in getting rid of the buggy `Tunresolved` type.

Next step: remove most cases by adding them to simplify_subtype

Reviewed By: andrewjkennedy

Differential Revision: D8222415

fbshipit-source-id: f95029512ff06e1a331500f6b65eeac9ee35b2d6