allow refinement of shape type using nullity Shapes::idx
Summary:
Refine the type of shapes in cases like
```
if (Shapes::idx($myshape, 'myfield') !== null) {
  // ...
}
```
This means that the shape now has the field `myfield`, and that the type for this field is not nullable.

We stay quite permissive: we add the field to the shape type regardless of whether this field can be here at all. Errors will anyway be raised elsewhere when typechecking the call to `Shapes::idx`. This allows for more useful typechecking of incomplete code (code in the process of being written).

Also refactor `refine_lvalue_type`.

Reviewed By: Wilfred

Differential Revision: D8222849

fbshipit-source-id: 232715367fb02360e54c92dc7730b11be396735e