Tracepoints for String copy-on-write events
Summary:
User space tracepoints are attractive because they can be given stable names
that we document, they're inserted into carefully curated locations in our runtime,
and compile to a NOP when not activated by tracing tools.

Instrument string modifications with probes for copy-on-write (cow) when we have
inspected the ref count and triggered a copy, or in-place mutation (mut) when we
modify a string in place after checking its ref count.

* No logging for "private string" mutations that don't check the ref-count, such as
from StringBuilder.
* No logging for other kinds of string copies. The can be added if they are interesting,
but we should take care to avoid double-counting copies.

Reviewed By: swtaarrs

Differential Revision: D9133241

fbshipit-source-id: b3d10659697986c14bd65148fbf656bbb02ae467