Fix lambda typechecking
Summary: The Hack typechecker does flow typechecking and as a result, updates the `env` as it is typechecking. We were originally updating the env to have `inside_ppl_class = false` when we were typechecking lambdas and anonymous functions. We were not changing the env to return to the original state after typechecking the anonymous function. This diff changes the typechecking env update to only affect the typechecking for the lambdas and anonymous functions.

Reviewed By: jamesjwu

Differential Revision: D9730635

fbshipit-source-id: 04540c3d8961f60a431dfd863056ba655e745a04