Add GC worker wall time pcts to profiling table
Summary:
On workers we measure how much wall time is spent in GC while running a job. This is cool to show in the profiling table!

Since it overlaps with the run time, we do need to subtract GC time from the run time to make the percentages add up to 100%

If it's not obvious, `GC m` is minor gc and `GC M` is major gc

Reviewed By: samwgoldman

Differential Revision: D9826945

fbshipit-source-id: ce4d31c8c5b20f17f7b563aa175ee1d10e07b78a