Make gen-type-scanners work with .debug_types
Summary:
This basically switches from using Dwarf_Off to using a new GlobalOff
type, which is an offset, and a flag. The flag indicates whether the
offset is in the .debug_info section (the old default), or the
.debug_types section.

This allows getAttributeValueRef to chase ref_sig8's back to their
origin in the .debug_types section. This breaks some assumptions (eg
that all references in a cu are to things within the cu) so I chased
those down.

Reviewed By: ricklavoie

Differential Revision: D9753534

fbshipit-source-id: 1598ad47633a4ae9d35d66194f2063110e813def