fb_compact_serialize: hack array tests
Summary:
add a hackarray version with the current output of running it with hack arrays enabled.
i make no assertion that any of those values are the correct output, but it is designed to surface changes to behavior, if they are made.

the existing test output does not change. in D9837094 i update the common include file with additional tests and more verbose output.

Reviewed By: ricklavoie

Differential Revision: D9833060

fbshipit-source-id: dc4af7a5337e9879ae9f307addf4ef3b6075439c