Remove get_bazooka
Summary:
The name `Typing_deps.get_bazooka` implies that this function somehow returns a larger set of dependencies than `Typing_deps.get_ideps`, in an attempt to be sure that we are rechecking everything which needs to be rechecked by discarding some of the granularity in our dependency tracking and checking some things which may not need rechecking.

However, in practice, `get_bazooka` is never called in such a way that it behaves any differently from `get_ideps`. The behavior of `get_bazooka` is identical to `get_ideps` for toplevel definitions (such as `Dep.Fun` or `Dep.Class`), and it is only ever invoked with a `Dep.variant` representing a toplevel definition.

I propose that we change all callsites of `get_bazooka` which do not rely on `get_bazooka`'s special behavior to just use `get_ideps` instead, since the use of `get_bazooka` only serves to make it less clear what is actually happening (i.e., nothing). Since no callsites currently rely on `get_bazooka`'s special behavior, I also propose that we do away with it altogether (but perhaps there is some upcoming use case for which we should keep it).

Reviewed By: dabek

Differential Revision: D9777000

fbshipit-source-id: 01f715a3869ae334f5cf3b3adfd2df1e5f2a7278