Survive EINTR while invoking external compiler.
When using the Xenon sampling profiler, HHVM will periodically thwack
itself with SIGVTALRM, causing blocking system calls to fail with EINTR.
This included reads over the pipe to the HackC child compiler, which
led to runtime failures.

With this patch we are able to run for long periods of time with the
Xenon profiler on, and hh_single_compile singly compiling its little
heart out. Sweet.