Convert {Is/As}TypeStruct to take the type structure from the stack
Summary:
This diff modifies `IsTypeStruct` and `AsTypeStruct` to take the type structure argument from the stack rather than as immediates. In doing so, it adds the `C` suffix to the bytecode to denote that type structure is taken as a `Cell` and also introduces to Resolve/DontResolve options to the bytecodes. Currently the regular is/as expressions are using the Resolve version and the reified generic is/as versions are using DontResolve version as `CombineAndResolveTypeStruct` instruction already does the resolution.

In the future, we can further optimize the Resolve version to change into DontResolve when we know that the type structure needs no resolution.

These new versions of the instructions allow us to do

```
function f<reified T>(mixed $x) {
  $x is T;
}
```
The previous versions of the instructions took the type structure as an immediate however with reified generics we do not have access to the type structure as an immediate. Hence, this diff adds the possibility to take the type structure from the stack.

Reviewed By: swtaarrs

Differential Revision: D9305649

fbshipit-source-id: 8a8ac02d549a48c7da67afafe36d5099c3e22bee