Test/fix similar files (file hashes during init)
Summary:
Similar files (D5972889) are an optimization that allows us not to fan-out from files that change in insignificant way, even if we don't have the previous versions of declarations (they use hash of previous version of AST with function bodies removed).

It was accidentally made less effective in D5360294, where doc blocks became part of ast.ml structure that we hash.

Reviewed By: jamesjwu

Differential Revision: D9793220

fbshipit-source-id: e84ed092bce6719a1623fb09f60267ff54c56e6d