Make clsref bigger
Summary:
This diff converts clsref from `LowPtr<Class>` to `LowPtr<Class>` and `ArrayData*` in order to prepare the codebase for carring the per instantiation reified generics of the class along with the class pointer.

Each reification of a class has a unique name and we need to be able to associate class pointer and the reified generics in `FPushCtor`. Another approach we tried as to have fpushctor pop additional argument from the stack but this does not work when classname is passed from a different function.

Reviewed By: jano

Differential Revision: D9772750

fbshipit-source-id: 43820b54c1ecfc4c99180c55de065a77de92fca0