Approximate union of tuples by a tuple of unions
Summary:
At the moment, the union of `(t0, ..., tn)` and `(t0', ..., tn')` is computed as `(t0, ..., tn) | (t0', ..., tn')` unless the two tuple contain equivalent types.  This leads to an exponential blowup if we have code like
```
if ($condition0) {
  $t[0] = $x0;
} else if ($condition1) {
  $t[1] = $x1;
} else if ...
```
inside a loop and `$t : (t0, ..., tn)`, `$x0 : t0'`, `$x1 : t1'` etc.  After a single iteration the type of `$t` will be a union `(t0', t1, ..., tn) | (t0, t1', ..., tn) | ... (t0, t1, ..., tn') | (t0, t1, t2, ..., tn)`, and after two iterations the type of `$t` will be a union containing an exponential number of tuples corresponding to all possible ways of combining `ti` and `ti'` in each component `i` (and some tuples may appear multiple times).  Depending on the lengths of the dependency chains of variables inside the body of the loop, we may need to make more than two iterations, which will lead to an even bigger type for `$t`.

To work around this issue, let us approximate the union of `(t0, ..., tn)` and `(t0', ..., tn')` by `(t0 | t0', ..., tn | tn')`.  Note that we already do this for tuple-like arrays as well as all covariant types including Hack arrays.

Reviewed By: kmeht

Differential Revision: D9851948

fbshipit-source-id: 58ee7f010b6938e7ee60355c1dfc846208fccb69