Don't bother sorting the functions again
Summary:
This function relied on the ordering of booleans, which isn't immediately obvious.

We're passing hack functions in the correct order anyway, so remove the sort. Closure_convert.convert_toplevel_prog ensures that top-level functions come last, so we don't need to sort again.

Since List.sort in Core_kernel is currently stable, this isn't actually changing the order.

Reviewed By: vassilmladenov

Differential Revision: D9831386

fbshipit-source-id: f2071bd2e0b1df93eee20474851eb37c27d434fb