Exclude top gCPU consumers from mode/opt Clang (NFC)
Summary:
**This diff will not be landed until September 17th 2018 at 1PM ET.**
For more context, please read a group post that will be published in the
fbcode group on September 10th.

T32219529 tracks flipping mode/opt builds in fbcode from using GCC, to using
Clang instead. The work is described in detail here:
https://fb.facebook.com/groups/fbcode/permalink/1825595014143996/

As explained above, the migration to using Clang is meant to only impact the
"long tail" of fbcode services that consume a great deal of build time, but less
than 1% of gCPU. This diff explicitly excludes the top gCPU consumers from the
migration.

This change on its own results in no functional change at all -- all of these
services already use GCC in mode/opt. But these allow a change to mode/opt's
compiler to build and test successfully.

Reviewed By: igorsugak

Differential Revision: D9411889

fbshipit-source-id: ca856e81de4a16ab0a2e643d77fe6a2f7733e722