Re-land "[fbcode] Exclude top gCPU consumers from mode/opt Clang (NFC)"
Summary:
Depends on D9931181.

The initial attempt to land this change, D9411889, caused Buck assertions due
to the interactions detailed in T34130018. D9931181 addresses that problem and
now this diff should, in fact, be NFC (no functional change).

The timing of this migration was announced to fbcode as Sep 18. I'd like to try
to stick to this timeline, so I'll land this diff as soon as it is accepted
and Sandcastle reports green.

Reviewed By: yfeldblum

Differential Revision: D9931656

fbshipit-source-id: e7f03d5b40af74a25d810dc935f70ab311d7782f