Don't let keysets break back-references in serialized data.
Summary:
We were incrementing the value count in the serializer for each value in a keyset, but keyset values are also keys. It's not possible to have back-references to keys in any container, and the unserializer was not counting them, so back-references to objects/refs that were first seen after a non-empty keyset would not round-trip: https://3v4l.org/n11oa
Fix this by no longer counting keyset values when serializing. This way we don't change the behavior of unserializing anything we've already serialized and stored somewhere, things serialized from this point forward will round-trip correctly, and we don't have to have special code in the unserializer to count (and then reject back-references to) keyset keys (we have such code for keys in Map and Set collections, because unfortunately we also count their keys).

Reviewed By: ricklavoie

Differential Revision: D9931268

fbshipit-source-id: 52a3ab9e7643e76ec2ecba65ea9a2751cc84131d