Typechecker process can query Mercurial on hg.update state leave events to log mergebase changes
Summary:
To make smarter decisions about restarting vs. proceeding incrementally we would like to answer about every file change "was it a part of a rebase or a local change by a user". There is a meeting this week with Watchman team on how we can do it in a fast and reliable way (possibly by them providing some additional APIs), but in the meantime I would like to get something quick to roll out for a week to small percentage of users and measure impact of this entire endeavor.

The main problem with this approach is that we are now blocked on Mercurial in more places than before, and will a) regress experience when Mercurial is having problems b) we will get the blame for Mercurial problems, because we are bad at surfacing underlying hh_server slowness reasons to the users

Reviewed By: pittsw

Differential Revision: D9947630

fbshipit-source-id: 389f7e1c0e4547b65dc09bbcb60dfbaa82e56aa6