Optimize as expressions to run is expression check first
Summary:
This is a very straight forward optimization that converts
```
$x as int
```
into
```
($x is int) ? $x : ($x as int)
```
while avoiding the double evaluation. This makes is so that the fast pass for `$x as int` is very fast but it also comes with the disadvantage of the slow path being slower since now we have an additional `is` check.

In addition, we are resolving the typestructure twice which can also be optimized away with the new Resolve/DontResolve bytecode extensions but it will require tiny bit more work on the compiler, hence this diff deters it.

Reviewed By: kmeht

Differential Revision: D9947989

fbshipit-source-id: d946447c83e29273077adf53008442784d30954a