Fix bugs lowering cmpbm and cmpwm
Summary:
They were always turned into unsigned compares. In practice, we only
generate cmpbm to check rds generation numbers, which is an equality
check; and we only use cmpwm for checking classVecLen, which *is* an
unsigned compare. But we also opportunistically created them in
vasm-simplify (which was another bug - its never profitable to do so,
and it could in theory break because simplify runs both before and
after lowering) so we *could* end up doing the wrong thing.

After this diff we generate the same code when the compares really are
unsigned, and we will generate worse, but correct code if we ever use
these instructions for anything else.

I also noticed that we can make some of the lowering a bit cheaper;
vasm_lower should just do nothing if it doesn't want to change the
instruction, rather than re-emitting a copy of it.

Reviewed By: ottoni

Differential Revision: D9951959

fbshipit-source-id: e74a9b039bf3d1aeedeefdfbdabbe8718ebdd533