Fix end offset for HH_FIXME pos
Summary:
The FFP is currently putting the end_offset of a `HH_FIXME` as the start offset of the node the `HH_FIXME` relates to. However, if we have code where there's multiple `HH_FIXME`s, e.g.

```
/* HH_FIXME[4000] This no longer applies */
/* HH_FIXME[4001] This still applies */
if (some_expr) { ... }
```

and we want to remove `HH_FIXME[4000]` using the start and end offsets, we'll strip off both `HH_FIXME`s instead of the first one.

This diff changes the end_offset to equal the end_offset of the `HH_FIXME`. As the end_offset includes the last character, we also need to update `clientRefactor` and add 1 to `char_end`.

Reviewed By: Wilfred

Differential Revision: D9934448

fbshipit-source-id: 9a52a18051f277782126f6fcde5c6e3e74a8719a