Laxer typing rules for append and assignment to collections with reference semantics
Summary:
This diff aims to fix the inadequacies of the typing rules for append and array assignment that I noticed while working on D9677742.

Consider the following example: `$p : Pair<t1, Vector<t2>>`, `$x : t2`, and we are trying to append `$x` to `$p[1]`:
```
$p[1][] = $x;
```
This should type-check: `$p[1]` gives us a reference to the `Vector`, which we destructively mutate preserving its object identity. However, this code doesn't type-check at the moment because our rules suggest that we first need to check if `$x` can be appended to `Vector<t2>` and then if `Vector<t2>` (the type arising from the append) can be assigned back to `$p`.  Because `Pair`s are immutable, a direct assignment to a component of a pair is rejected.  Note that HHVM happily executes the above code.  On the other hand, if `$p : Pair<t1, vec<t2>>`, then the above append code raises an `InvalidOperationException`, and hence should also be rejected by the type-checker.

It seems that we need to distinguish between reference and non-reference types.  Reference types are `Vector`s, `Map`s, and `Set`s, as well as arbitrary unions thereof.  Non-reference types are PHP and Hack arrays and unions containing non-reference types.  For reference type, we only check if the type supports appends / assignments of values of a given type, but don't check if the resulting type (which will always be the same as the initial type) is assignable back to the lvalue.

Reviewed By: CatherineGasnier

Differential Revision: D9916492

fbshipit-source-id: bec6fd7e34fe47c580f5a59f2b58f5f43662ca38