Create RDS_LOCAL macros: RDS equivalent of THREAD_LOCAL
Summary:
This diff adds RDS_LOCAL macros.  They are an RDS equivalent of THREAD_LOCALs.  Most of our current THREAD_LOCALs should logically be tied to a request.  Currently threads are mapped 1 to 1 to requests, so this works great, but in the future this may not be the case.  This makes it possible to store request data in our RDS segment using a similar API to THREAD_LOCAL.

This also adds macros for hot rds locals.  These are stored flat in thread local storage to allow for quick access using %fs offset addressing.  To make them effectively RDS local, the memory will be copied into and out of RDS every time we switch a threads active RDS.

As a part of this diff, RequestLocals using the REQUEST_LOCAL macros are converted to use an RDS_LOCAL backing.

Reviewed By: jjgriego

Differential Revision: D9242916

fbshipit-source-id: a64a355342b89e49434af6a2475b2088d348ab51