Refactor / simplify prechecked files init
Summary:
The goal of this diff is to move the place where we expand `local_dirty_files` to happen in `ServerInit` during server startup, and not in `ServerPrecheckedFiles` after initial round of typechecking is done.

The reasons why it was not done in the first place were eliminated in D9322210	and D9321588.

This allows to pull out more parts of `update_after_recheck` and `update_after_local_changes` into a shared common function.

Reviewed By: arxanas

Differential Revision: D9977391

fbshipit-source-id: 7c0a60e0ed7ca5b87f547707d70e0e1a5df337eb