Loading saved states with errors
Summary:
Adding the ability to load a saved state that has errors. Note that the format of the saved state changes, even if we don't allow errors.
Saving the errors together with the files' info (the naming table).
Also loading saved states that may contain a set of files with errors.

Reviewed By: dabek

Differential Revision: D9932749

fbshipit-source-id: fe1148d32ce255ceafda75e99979b206505b783c